Class Online assignment to individual instructors update
Add new assignments Structure. Sections in assignments object no longer array. It's an object with each section as the Key with class: and online: keys to boolean values.

Update toggleSection function. Toggles which parts of a section (class, online or both) are assigned to a specific instructor for a specific (course, semester, section), enforcing exclusivity (at most on instructor owns class and at most one owns online for that section).